🚝👷 Checkpoint ./komutlar/durum-süre.js:26490528/195

+++ komutlar/durum-süre.js
@@ -28,6 +28,7 @@ message.react('✅')
 }
   
 if (status === "kapat"){
+if (!data2) return message.channel.send(`${basarisiz} ${message.author}, Durum zamanı zaten kapalı.`).then(x => x.delete({timeout: 5000}))
 await db.delete(`durum_süresi`)
 if (!data){
 message.channel.send(`${basari} ${message.author}, Durum zamanınız kapatıldı.`)
